test: move multipleCancels into an isolated SharedTimerRefCountTest
multipleCancels asserts the process-global SharedTimer ref count. That
only holds when no other test mutates the timer concurrently and when no
unrelated connection already holds a reference. Neither is guaranteed
once the query_timeout=15 matrix axis applies a JVM-wide default
queryTimeout: the setUp connection then holds a reference, so the count
starts at 1 and the old "expected 0" assertion failed.

Extract the test into its own @Isolated class so no concurrent test can
perturb the global count, and assert deltas from a baseline captured at
the start rather than an absolute 0/1/0. This mirrors the existing
@Isolated convention (e.g. CursorFetchSqlTransactionTest for pg_cursors).
